split legacy methods to IrcLegacySubscriber
this makes more distinct what will be dropped in the future
