+++ src/Event/IrcSubscriber.php
@@ -15,10 +15,8 @@ namespace Eventum\Event;
 
 use Date_Helper;
 use DB_Helper;
-use Group;
 use Issue;
-use Notification;
-use Reminder;
+use Setup;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Workflow;
@@ -32,124 +30,20 @@ class IrcSubscriber implements EventSubscriberInterface
     {
         return [
             SystemEvents::IRC_NOTIFY => 'notifyIRC',
-            SystemEvents::IRC_NOTIFY_BLOCKED_MESSAGE => 'notifyBlockedMessage',
-            SystemEvents::NOTIFY_ISSUE_CREATED => 'notifyIssueCreated',
-            SystemEvents::REMINDER_ACTION_PERFORM => 'reminderAction',
         ];
     }
 
     /**
-     * Method used to send an IRC notification about a blocked email that was
-     * saved into an internal note.
-     *
-     * @deprecated implement the logic in your own Subscriber
-     */
-    public function notifyBlockedMessage(GenericEvent $event)
-    {
-        if ($event['irc_legacy_handled']) {
-            return;
-        }
-
-        $issue_id = $event['issue_id'];
-        $from = $event['from'];
-
-        $notice = "Issue #$issue_id updated (";
-        // also add information about the assignee, if any
-        $assignment = Issue::getAssignedUsers($issue_id);
-        if (count($assignment) > 0) {
-            $notice .= 'Assignment: ' . implode(', ', $assignment) . '; ';
-        }
-        $notice .= "BLOCKED email from '$from')";
-
-        $prj_id = Issue::getProjectID($issue_id);
-        Notification::notifyIRC($prj_id, $notice, $issue_id);
-    }
-
-    /**
-     * @param GenericEvent $event
-     * @deprecated implement the logic in your own Subscriber
-     */
-    public function reminderAction(GenericEvent $event)
-    {
-        if ($event['irc_legacy_handled']) {
-            return;
-        }
-
-        $issue_id = $event['issue_id'];
-        $action = $event['action'];
-
-        // alert IRC if needed
-        if (!$action['rma_alert_irc']) {
-            return;
-        }
-
-        $irc_notice = "Issue #$issue_id (";
-        if (!empty($data['pri_title'])) {
-            $irc_notice .= 'Priority: ' . $data['pri_title'];
-        }
-        if (!empty($data['sev_title'])) {
-            $irc_notice .= 'Severity: ' . $data['sev_title'];
-        }
-        // also add information about the assignee, if any
-        $assignment = Issue::getAssignedUsers($issue_id);
-        if (count($assignment) > 0) {
-            $irc_notice .= '; Assignment: ' . implode(', ', $assignment);
-        }
-        if (!empty($data['iss_grp_id'])) {
-            $irc_notice .= '; Group: ' . Group::getName($data['iss_grp_id']);
-        }
-        $irc_notice .= "), Reminder action '" . $action['rma_title'] . "' was just triggered; " . $action['rma_boilerplate'];
-
-        $prj_id = Issue::getProjectID($issue_id);
-        Notification::notifyIRC($prj_id, $irc_notice, $issue_id, false, APP_EVENTUM_IRC_CATEGORY_REMINDER);
-    }
-
-    /**
-     * Notify new issue to irc channel
+     * Save event details to irc_notice table.
      *
      * @param GenericEvent $event
-     * @deprecated implement the logic in your own Subscriber
      */
-    public function notifyIssueCreated(GenericEvent $event)
-    {
-        if ($event['irc_legacy_handled']) {
-            return;
-        }
-
-        $issue_id = $event['issue_id'];
-        $prj_id = $event['prj_id'];
-        $data = $event['data'];
-
-        $irc_notice = "New Issue #$issue_id (";
-        $quarantine = Issue::getQuarantineInfo($issue_id);
-        if ($quarantine) {
-            $irc_notice .= 'Quarantined; ';
-        }
-
-        $irc_notice .= 'Priority: ' . $data['pri_title'];
-
-        // also add information about the assignee, if any
-        $assignment = Issue::getAssignedUsers($issue_id);
-        if (count($assignment) > 0) {
-            $irc_notice .= '; Assignment: ' . implode(', ', $assignment);
-        }
-
-        if (!empty($data['iss_grp_id'])) {
-            $irc_notice .= '; Group: ' . Group::getName($data['iss_grp_id']);
-        }
-        $irc_notice .= '), ';
-
-        if (isset($data['customer'])) {
-            $irc_notice .= $data['customer']['name'] . ', ';
-        }
-
-        $irc_notice .= $data['iss_summary'];
-
-        Notification::notifyIRC($prj_id, $irc_notice, $issue_id, false, false, 'new_issue');
-    }
-
     public function notifyIRC(GenericEvent $event)
     {
+        if (!$this->notificationEnabled()) {
+            return;
+        }
+
         $category = $event['category'];
         $notice = Workflow::formatIRCMessage(
             $event['prj_id'], $event['notice'], $event['issue_id'],
@@ -179,5 +73,12 @@ class IrcSubscriber implements EventSubscriberInterface
 
         $stmt = 'INSERT INTO `irc_notice` SET ' . DB_Helper::buildSet($params);
         DB_Helper::getInstance()->query($stmt, $params);
+    }
+
+    private function notificationEnabled()
+    {
+        $setup = Setup::get();
+
+        return $setup['irc_notification'] === 'enabled';
     }
 }
